requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-multipart
pypdf
openai
sqlalchemy
psycopg[binary]
redis
rq
qdrant-client==1.9.1
python-telegram-bot==21.11.1
